SheetScoutv2
A second iteration of the sheet scouter, involved some reformatting of the structure. Each column in the spreadsheet is represented by a data class in the sequence class. 

**Data.java - data classes representing each type of data with its on method of processing and analyzing the data

**Trigger.java - takes a set of data and returns a boolean if the trigger is true i.e. a comparison trigger if any of set [x] > y return true

Sequence - Determine what is to be done for each column of the output table, average the values, use a trigger value etc, used in the sheetsequencer

Parameter parser- parse set of parameters in the following format 'NAME ATTRIBUTE=VALUE ...'

Project was abandoned again due to some structural issues and rewritten