SheetScoutv3
Similar to v2 in that it uses data to store team data and triggers. Redesigned so that it is simpler, more modifiable, with sheet superclass, a planned sequence editor.

Sequence - Order of columns and the data type for each column
Team - Object that stores the data for each team
Sheet - a generic object for spreadsheets, i.e. googlesheets, excel sheet or otherwise

This project is a current WIP